+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - FEATURE kumo now has the visualize cmd. Req. dot installation (#136).
 - FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX ramuda rollback to previous version.
 - FIX [kumo] Parameter diffing does not work for aws coma-seperated inputs (#77).
 
 ## [0.0.57] - 2016-09-23